requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
asyncpg
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiohttp
yfinance
pandas
numpy
loguru
python-dotenv
schedule
